--- packages/client/src/components/business/App.tsx
@@ -1,16 +1,0 @@
-import react from 'react'
-import './App.css'
-import { Routes, Route, BrowserRouter } from 'react-router-dom'
-import { SignIn, SignUp, PageNotFound } from './pages'
-
-export const App = () => {
-  return (
-    <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<SignIn />} />
-        <Route path="/sign-up" element={<SignUp />} />
-        <Route path="*" element={<PageNotFound />} />
-      </Routes>
-    </BrowserRouter>
-  )
-}
