+++ packages/client/src/components/business/App/StyledComponents.tsx
@@ -1,0 +1,19 @@
+import styled from 'styled-components'
+import v from '../../../Variables.module.scss'
+
+const Wrap = styled.div`
+  box-sizing: border-box;
+  font-family: ${v.ff};
+  -moz-osx-font-smoothing: grayscale;
+  -webkit-font-smoothing: antialiased;
+  -webkit-text-size-adjust: 100%;
+  -ms-text-size-adjust: 100%;
+  -moz-text-size-adjust: 100%;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+`
+
+export { Wrap }
