+++ packages/client/src/components/business/App/index.ts
@@ -1,0 +1,1 @@
+export { App } from './App'
